Move photography related components into one folder
