+++ src/components/photography/Hexagon.js
