+++ src/components/photography/More.js
