+++ src/components/photography/TopNews.css
